add adminka, tests

+++ web_test/src/main/java/ru/webtest/springbootweb_test/repositories/QuestionRepository.java
@@ -4,8 +4,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.webtest.springbootweb_test.entitys.Question;
 
+import java.util.List;
+import java.util.Set;
+
 @Repository
 public interface QuestionRepository extends JpaRepository<Question,Long> {
-    Question[] findQuestionsByParent(int parent);
+  Question[] findQuestionsByParent(long parent);
+
 
 }
